% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_functions.R
\name{accessSPCDB}
\alias{accessSPCDB}
\title{Access database for spacy parses.}
\usage{
accessSPCDB(workingfolder)
}
\arguments{
\item{workingfolder}{folder to save file in.}

\item{workingfolder}{folder to save file in.}
}
\value{
will load a database names spacyframe.db

will create a database names spacyframe.db
}
\description{
This command creates a sqlite database which parse information is saved in.

This command creates a sqlite database which parse information is saved in.
}
\examples{
accessSPCDB()
Create an access database for spacy parses.

accessSPCDB()
}
\seealso{
the entire dplyr package

the entire dplyr package
}
