% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicFunctions.R
\name{AnnotateVerbsTopicJoin}
\alias{AnnotateVerbsTopicJoin}
\title{Keep only those sentences with desired verbs.}
\usage{
AnnotateVerbsTopicJoin(WT, PROCESSED, OUT, ANNOTATELIST, SENTENCEFRAME,
  toptopics)
}
\arguments{
\item{WT}{a ilst of verbs}

\item{PROCESSED}{an stm processed object.}
}
\value{
OUT an stm prepped object.

ANNOTATELIST openNLP annotation object

SENTENCEFRAME SentFrame object returned by PreTopicFrame

toptopics TopTopics extracted from an stm topic model.
}
\description{
This function parses sentences only keeping those sentences that include a verb that is in the list which is specified by the user. More verbs may be kept if they occur in the sentence alongside the desired verb.
}
\examples{
AnnotatesLarge<-AnnotateVerbsTopicJoin(allwords,BASE_INPUT$processed,BASE_INPUT$out,BASE_INPUT$Annotations,BASE_INPUT$SentFrame,BASE_INPUT$top.topics)
}
\seealso{
\code{\link{stm}}
}

