% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{assocPrettyOneStep}
\alias{assocPrettyOneStep}
\title{Makes a pretty word association table all in one step.}
\usage{
assocPrettyOneStep(wordlist, termDocumentMatrix, corpus, corrVal = 0.8)
}
\arguments{
\item{wordlist}{list of words for association}

\item{termDocumentMatrix}{tdm of corpus}

\item{corpus}{the corpus the words are from}

\item{corrVal}{the correlation value to report at or above}
}
\value{
a word association table
}
\description{
This function takes a wordlist, tdm, coprus, and settable correlation value and returns a datatable.
}
\examples{
assocPrettyOneStep(c("research","find"),termDocumentMatrix(corpus),corpus1,.8)
}
\seealso{
\code{\link{findAssoc}}
}

