% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{assocPTable}
\alias{assocPTable}
\title{Makes a pretty word association table .}
\usage{
assocPTable(assoctable, corpus, ngram = FALSE)
}
\arguments{
\item{assoctable}{rough word association table}

\item{corpus}{word corpus for table}
}
\value{
a cleaned tm corpus
}
\description{
This function takes a findAssoc object and returns a datatable. assocPrettyOneStep is wraps this to allow everything in one step.
}
\examples{
assocPTable(findAssoc(corpus1,wordlist,corrVal),corpus)
}
\seealso{
\code{\link{tm_map}}
}

