% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicFunctions.R
\name{buildcliff}
\alias{buildcliff}
\title{These commands start and run a cliff-docker server.}
\usage{
buildcliff()
}
\description{
These commands collectively build, ensure a server is working, and stop a running docker server for gecoding. Once you have run buildcliff and startcliff, you can run the PredictCountryByDoc command to predict which countries documents correspond to. For these commands, we need to have the opportunity id labels as OpID in the SentFrame part of BASE_INPUT
}

