% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{checkForTika}
\alias{checkForTika}
\title{Check to see if Tika is available.}
\usage{
checkForTika(directory = getwd())
}
\arguments{
\item{directory}{directory to look for tika in.}
}
\value{
If false will download tika to working directory.
}
\description{
The check for tika command will install tika to the current working directory.
}
\examples{
checkForTika()
}
\seealso{
\code{\link{http://apache.claz.org/tika}}
}

