% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_functions.R
\name{conligraph}
\alias{conligraph}
\title{It makes a network graph of words. First part of list is a data frame showing connections, Second part I believe is an actual igraph object.}
\usage{
conligraph(parseconll)
}
\arguments{
\item{parseconnll}{spacy parsed frame}

\item{tmidf}{tfidf frame}

\item{scydb}{db with spacy in it}
}
\description{
This command makes a network graph of words from a spacy output frame.
}
