% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{copyDir}
\alias{copyDir}
\title{Creates directory and copies from old to new.}
\usage{
copyDir(from, to)
}
\arguments{
\item{from}{Old directory}

\item{to}{New directory}
}
\value{
NA, but he file will be relocated
}
\description{
This function is just a basic wrapper to copy and move folders recursively in linux. Should probably just use file.copy instead.
}
\examples{
copyDir("../doc","../newdocs")
}
\seealso{
\code{\link{file.copy}}
}

