% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formFunctions.R
\name{datapicker}
\alias{datapicker}
\title{Try to auto-extract form information.}
\usage{
datapicker(file_list, multiple = TRUE, graphics = FALSE)
}
\arguments{
\item{file_list}{list of word document files}

\item{multiple}{allow selecting multiple words}

\item{graphics}{graphical or boring interface.}
}
\description{
a method for viewing the content of a word
}

