% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epar.example-data.R
\docType{data}
\name{demo.docs}
\alias{demo.docs}
\title{Example EPAR data}
\format{An object of class \code{"data.frame"}; see \code{\link[qtl]{data.frame}}.}
\source{
\href{evans.uw.edu}{Evans School}
}
\usage{
data(demo.docs)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\examples{
data(demo.docs)
}
\references{
evans.uw.edu
(\href{evans.uw.edu}{Evans School})
}
\keyword{datasets}

