% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{doc_clean_process}
\alias{doc_clean_process}
\title{Cleans documents performing many common tasks .}
\usage{
doc_clean_process(corpusname)
}
\arguments{
\item{corpus}{corpus to clean}
}
\value{
a cleaned tm corpus
}
\description{
This function strips whitespace, removes stop words,removes punctuation,and stems documents.
}
\examples{
doc_clean_process(corpus1)
}
\seealso{
\code{\link{tm_map}}
}

