% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formFunctions.R
\name{docx_table_view}
\alias{docx_table_view}
\title{See xml table contents of a docx file in a viewer window.}
\usage{
docx_table_view(file, export_frame = FALSE, showView = T)
}
\arguments{
\item{file}{file to view xml table contents of}

\item{export_frame}{should the data frame be saved to output?}
}
\value{
view of file
}
\description{
a method for viewing the content of a word document via parsing the xml file to a data frame.
}

