% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{example_documents}
\alias{example_documents}
\title{Build Example Data.}
\usage{
example_documents()
}
\value{
a folder of documents
}
\description{
This function creates a folder if documents in the working directory.
}
\examples{
example_documents()
}

