% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formFunctions.R
\name{formcluster}
\alias{formcluster}
\title{Cluster forms based on xml content.}
\usage{
formcluster(file_list)
}
\arguments{
\item{file_list}{list of word document files}
}
\description{
a method for viewing the content of a word
}

