% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{how_do_i}
\alias{how_do_i}
\title{Look up how to do something in R}
\usage{
how_do_i(theproblem)
}
\arguments{
\item{theproblem}{the problem.}
}
\description{
A joke command to look up a method in R.
}

