% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{how_do_i_open}
\alias{how_do_i_open}
\title{Look up how to open a file.}
\usage{
how_do_i_open(ending)
}
\arguments{
\item{ending}{the ending of the file in question.}
}
\description{
Look up how to open a file.
}

