% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_functions.R
\name{IDframes}
\alias{IDframes}
\title{Select relevant frames to keep}
\usage{
IDframes(VERBWORD, parsecnnl)
}
\arguments{
\item{VERBWORD}{regular expression for verb you would like to generally keep}

\item{parsecnnl}{table resulting from spacy parse}
}
\description{
This command uses framenet to identify verb frames
}
\examples{
IDframes("[cause|change],collect(tbl(scydb,"parses_uw"),n=Inf)")
}
