% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicFunctions.R
\name{idtopics}
\alias{idtopics}
\title{Identify topics with highest term probability}
\usage{
idtopics(STMOBJ, TERM, N)
}
\arguments{
\item{STMOBJ}{stm object}

\item{TERM}{word in character format}
}
\value{
N number of topics to return
}
\description{
taking a topic model, this function returns the top topics for that word.
}
\examples{
idtopics(st1,"research",10)
}

