% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{interest_plot}
\alias{interest_plot}
\title{Creates ggplot of wordcounts!}
\usage{
interest_plot(wordlist, termDocumentMatrix, by.var = NULL, byvarname = "")
}
\arguments{
\item{wordlist}{list of words}

\item{termDocumentMatrix}{tm TermDocumentMatrix object}

\item{by.var}{can specify variable to group documents by}

\item{name}{plottable name of by.var}
}
\value{
a ggplot2 object
}
\description{
This function ggplot of wordcounts
}
\examples{
interest_plot("research",TermDocumentMatrix(corpus1),by.var=variable,"DocID")
}
\seealso{
\code{\link{ggplot2}}
}

