% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{interest_plot_bydoc}
\alias{interest_plot_bydoc}
\title{Creates ggplot of wordcounts by document}
\usage{
interest_plot_bydoc(wordlist, termDocumentMatrix)
}
\arguments{
\item{wordlist}{list of words}

\item{termDocumentMatrix}{tm TermDocumentMatrix object}
}
\value{
a ggplot2 object
}
\description{
This function ggplot of wordcounts by document id.
}
\examples{
interest_plot_bydoc("research",TermDocumentMatrix(corpus1))
}
\seealso{
\code{\link{ggplot2}}
}

