% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{interest_plot_bydoc_char}
\alias{interest_plot_bydoc_char}
\title{Creates ggplot of wordcounts by document, allowing external characteristics.}
\usage{
interest_plot_bydoc_char(wordlist, termDocumentMatrix, doccharacteristic)
}
\arguments{
\item{wordlist}{list of words}

\item{termDocumentMatrix}{tm TermDocumentMatrix object}

\item{doccharacteristic}{a vector the same length as tm specifying the trait you'd like to plot by}
}
\value{
a ggplot2 object
}
\description{
This function ggplot of wordcounts by document characteristic from an external list.
}
\examples{
interest_plot_bydoc("research",TermDocumentMatrix(corpus1),rep("bacon",length(corpus1)))
}
\seealso{
\code{\link{ggplot2}}
}

