% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicFunctions.R
\name{jgc}
\alias{jgc}
\title{Clean the java environment.}
\usage{
jgc()
}
\description{
Function cleans the java environment. This is a great command. Run it whenever you feel like it.
}
\examples{
jgc()
}
\seealso{
\code{\link{rJava}}
}

