% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_functions.R
\name{join_parse_tmidf}
\alias{join_parse_tmidf}
\title{Keep only sentences from relevant frames}
\usage{
join_parse_tmidf(parseconnll, scydb)
}
\arguments{
\item{parseconnll}{spacy parsed frame}

\item{scydb}{db with spacy in it}

\item{tmidf}{tfidf frame}
}
\description{
This command joins a parseconll and term matrix for a database.
}
