% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicFunctions.R
\name{makeadockercliff}
\alias{makeadockercliff}
\title{Run this function to install a docker server from within R}
\usage{
makeadockercliff()
}
\arguments{
\item{password}{you should type in your password 3 times.}
}
\value{
NA installs a docker server
}
\description{
This function calls out using the system command. make sure to clear your workspace and history if you use this. However, what a cool function, right?
}

