% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_functions.R
\name{maketermkeywords}
\alias{maketermkeywords}
\title{Create a tmidf matrix for keyword extraction}
\usage{
maketermkeywords(CORPUS_A, db)
}
\arguments{
\item{CORPUS_A}{corpus to analyze}

\item{db}{database to save tfidf to}
}
\value{
will create a database names spacyframe.db
}
\description{
This command creates a sqlite database which parse information is saved in.
}
\examples{
accessSPCDB()
}
\seealso{
the entire dplyr package
}
