% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{makeworking}
\alias{makeworking}
\title{Tests if save directory exists and creates new one if not}
\usage{
makeworking(dir)
}
\arguments{
\item{dir}{directory}
}
\description{
This function does the quite useful task of making a workingfolder object in the environment.
}

