% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formFunctions.R
\name{outcome_extractor}
\alias{outcome_extractor}
\title{Extract Outcomes from Gates proposal documents}
\usage{
outcome_extractor(file)
}
\arguments{
\item{file}{File to read in for scraping.}
}
\value{
a data frame, if only file exists, there was an error.
}
\description{
Extracts the outcome table from a Gates proposal document.
}
\examples{
outcome_extractor()
}

