% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicFunctions.R
\name{picktopics}
\alias{picktopics}
\title{Pick topics with suggested words}
\usage{
picktopics(STMOBJ, TERM, N)
}
\arguments{
\item{STMOBJ}{stm object}

\item{TERM}{word in character format}
}
\value{
N number of topics to return
}
\description{
Pulls similar words, expanding topic selection ability. Interactive.
}
\examples{
idtopics(st1,"research",10)
}

