% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicFunctions.R
\name{PreTopicFrame2}
\alias{PreTopicFrame2}
\title{Process a corpus into a topic model ready object, but do so while using the Syntaxnet Parser rather than the openNLP parser.}
\usage{
PreTopicFrame2(CORPUS_A, sample_num, syntaxnet = T, workingfolder,
  removeentities = T)
}
\arguments{
\item{CORPUS_A}{Document corpus}

\item{sample_num}{If sampling from documents, how many chunks to sample from each document? If 0, all chunks are kept.}

\item{syntaxnet}{Defaults to true but if it is false, this skips annotating for NLP and goes directly to text processing. This might be useful if you dont care about limiting verb-phrases in later steps and want to fit a topic model to all documents.}

\item{workingfolder}{folder where files are to be saved.}

\item{removeentities}{TRUE removes all proper nouns. FALSE keeps all proper nouns. "ONLY" will keep only proper nouns. This means you could fit a topic model to only proper nouns contained in documents}
}
\value{
SentFrame data frame with one row for each paragraph chunk

Annotations Rather than being stored in memory, annotations are stored in a sqllite database in the workingfolder in a file named textDB. This saves on memory. For each document, annotations are stored in a table named according to the document id. There are three additional tables, one of just propeer nouns called propers, one of nonproper nouns called nonprops, and one with everything called fulltable.

processed stm processed documents

out stm out documents for use in topic model
}
\description{
This function takes a corpus and creates a processed version of that corpus with entities removed for use in a topic model. Additionally it allows you to specify common entities to count across documents for use as a covariate in the topic model. The object it returns includes a frame of the text, an annotation object, a processed version of the corpus with stems and stopwords removed, and an out object which is the input object for fitting a topic model within the stm package.
}
\examples{
BASE_INPUT<-PreTopicFrame2(CORPUS_A,0,syntaxnet=F,workingfolder)
}
\seealso{
\code{\link{stm}}
}

