% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicFunctions.R
\name{processFolder}
\alias{processFolder}
\title{Process Documents in the Alchemy Folder}
\usage{
processFolder(workingfolder, PREPFRAME)
}
\arguments{
\item{workingfolder}{Folder for working in}

\item{PREPFRAME}{prepframe object}
}
\value{
data frame of processed sentences
}
\description{
This function parses the entries in the alchemy api folder and creates a dataframe
}

