% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicFunctions.R
\name{ProcessforAPI}
\alias{ProcessforAPI}
\title{Clean sentences for sending to alchemy API.}
\usage{
ProcessforAPI(PREPFRAME)
}
\arguments{
\item{PREPFRAME}{object returned by combinationframe or annotateframebytopicjoin}
}
\value{
a data frame
}
\description{
This function cleans the sentences for sending to alchemy API
}

