% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{processURI2}
\alias{processURI2}
\title{Worker function for processing URI info. From tm package.}
\usage{
processURI2(uri)
}
\arguments{
\item{uri}{the path to the uri to be processed.}
}
\value{
uri
}
\description{
Worker function for processing URI info. From tm package.
}
\seealso{
\code{\link{tm}}
}

