% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{read_docxtm}
\alias{read_docxtm}
\title{Reads docx files into tm corpus.}
\usage{
read_docxtm(file, skip = 0)
}
\arguments{
\item{file}{File to read. Should end in docx}

\item{skip}{Value is settable for debugging. Skips lines. Set at 0 for most applications.}
}
\value{
a plain text document corpus
}
\description{
This function reads a docx file into a tm corpus.
}
\examples{
read_docxtm("demodoc.docx",skip=0)
}
\seealso{
\code{\link{PlainTextDocument}}
}

