% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formFunctions.R
\name{readfiletab_gates}
\alias{readfiletab_gates}
\title{Read in a Gates proposal template.}
\usage{
readfiletab_gates(tfile)
}
\arguments{
\item{tfile}{File to read in for scraping table}
}
\value{
a data frame, if only file exists, there was an error.
}
\description{
Read the primary table from a Gates proposal word document, keeping column names.
}
\examples{
gates<-bind_rows(lapply(tfiles,readfiletab_gates))
}

