% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{readMails}
\alias{readMails}
\title{Read Emails into Corpus.}
\usage{
readMails(folder_in, newmailsdirectory)
}
\arguments{
\item{folder_in}{The path to the folder where the emails are located}

\item{newmailsdirectory}{The path to the folder where the processed emails will be stored.}
}
\value{
A tm Corpus of processed emails.
}
\description{
A tm Corpus of processed emails which can be combined with a corpus created by tm using the c() command. The command calls to msgconvert, and will only work if you have msgconvert installed.
}
\examples{
readMails("../doc","../procdocs")
}
\seealso{
\code{\link{http://www.matijs.net/software/msgconv/}}
}

