% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{readPDF2}
\alias{readPDF2}
\title{Function call to read pdf files into tm corpus.}
\usage{
readPDF2(engine = c("xpdf", "Rpoppler", "ghostscript", "Rcampdf", "custom",
  "pdftools"), control = list(info = NULL, text = NULL))
}
\arguments{
\item{engine}{which pdf library to use. see tm readPDF for more details}
}
\value{
engine for reading pdf
}
\description{
This function is a port of readpdf.
}
\examples{
readPDF2(engine="xpdf")(elem=list(uri=fname), language="en")
}
\seealso{
\code{\link{readPDF}}
}

