% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicFunctions.R
\name{RunAlchy}
\alias{RunAlchy}
\title{Function to run .R file returned by writerun_alch}
\usage{
RunAlchy(num)
}
\arguments{
\item{num}{row number to start function at (this is to allow for restarts in case an error occurs?)}
}
\value{
NA fills folder with alchemy api calls
}
\description{
This function runs the .R file returned by writerun_alch. It takes as an argument a number which corresponds to a row in the BASE_INPUT dataframe
}

