% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicFunctions.R
\name{runSTM}
\alias{runSTM}
\title{Run STM in a seperate R session.}
\usage{
runSTM(workingfolder)
}
\arguments{
\item{workingfolder}{the workingfolder where you are saving objects.}
}
\value{
writes file and starts new r process in background
}
\description{
this function will create .R file in the working folder which it then will call, resulting in fitting an stm model within a seperate R session. The output should be viewable in an Rout file.
}
\examples{
runSTM("../Research.Grants")
}
\seealso{
\code{\link{stm}}
}

