% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubMedviaRISmed.R
\name{searchPMIDs}
\alias{searchPMIDs}
\title{Return structured dataframe based on PMID info.}
\usage{
searchPMIDs(pmids)
}
\arguments{
\item{pmids}{character vector of pubmed ids}
}
\value{
a dataframe of info on requested publications
}
\description{
This function extracts pubmed data and returns a dataframe based on a pubmed id. More variables could be added by adding additional names and commands from the RISmed package
}
\examples{
example_documents()
}

