% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubMedviaRISmed.R
\name{searchWordsPubMED}
\alias{searchWordsPubMED}
\title{Return structured dataframe based on search terms.}
\usage{
searchWordsPubMED(searchterms, email, retmax = 25)
}
\arguments{
\item{searchterms}{a query to search}
}
\value{
a dataframe of info on requested publications
}
\description{
This function parses a search and returns a dataframe based on a pubmed id. More variables could be added by adding additional names and commands from the RISmed package
}
\examples{
test1<-searchWordsPubMED("'gender equity'+environment","ryscott5@uw.edu",retmax=10)
head(test1)
}

