% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_functions.R
\name{sentkeeper}
\alias{sentkeeper}
\title{Keep only sentences from relevant frames}
\usage{
sentkeeper(idframes, parsecnnl, database = T)
}
\arguments{
\item{parsecnnl}{table resulting from spacy parse}
}
\description{
This command uses the result of IDframes and keeps only sentences with relevant verb frames
}
