% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_functions.R
\name{SPCY_PreTopicFrame}
\alias{SPCY_PreTopicFrame}
\title{Parse sentences strings using SPACY as the parsing engine, and prep sentences for running through topic model}
\usage{
SPCY_PreTopicFrame(CORPUS_A, sample_num = 0, workingfolder,
  removeentities = T, dbexists = FALSE, spcy = T, syntaxnet = F)
}
\arguments{
\item{CORPUS_A}{corpus to analyze}

\item{sample_num}{number of rows to sample default is 0 which keeps all rows.}

\item{workingfolder}{folder to save file in.}

\item{spcy}{use spacy to parse (set to True usually)}

\item{syntaxnet}{use syntaxnet to parse (set to False usually)}

\item{removeentitites}{should named entities be dropped from topic model}

\item{dbexits}{whether the spacy or sqlnet database exists}
}
\value{
will create a database named spacyframe.db
}
\description{
This command creates a sqlite database which parse information is saved in, while also creating the proper frames for running a topic model.
}
\seealso{
the entire dplyr package
}
