% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{suggestkeywords}
\alias{suggestkeywords}
\title{Suggests keywords based on the conceptnet API}
\usage{
suggestkeywords(keyword, path_root = F)
}
\arguments{
\item{keyword}{word to search conceptnet for}

\item{path_root}{if true, can put in word path returned from previous calls otherwise defaults to keyword.}
}
\description{
Suggests keywords based on the conceptnet API
}
