% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{wfplots}
\alias{wfplots}
\title{Creates plots of word frequencies}
\usage{
wfplots(termDocumentMatrix, typePlot = 1, wordcount, minfreq = 5,
  shortendoc = FALSE, allCorpus = FALSE)
}
\arguments{
\item{termDocumentMatrix}{tm TermDocumentMatrix object}

\item{wordcount}{How many words to graph}

\item{minfreq}{minimum number of times a word should occur to be reported}

\item{shortendoc}{Should document names be shortened?}

\item{allCorpus}{For plot type 1, would you like the plot for all documents combined?}

\item{typeplot}{1=word on x, 2=variable on x}
}
\value{
a ggplot2 object.
}
\description{
This function creates bar graphs in ggplot. It can be customized by adding items to the returned object.
}
\examples{
wfplots(TermDocumentMatrix(corpus1),typePlot=1,5,minfreq=5,shortendoc=T)
}
\seealso{
\code{\link{ggplot}}
}

