% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{word_heatmap}
\alias{word_heatmap}
\title{TermDocumentMatrix to interactive heatmap.}
\usage{
word_heatmap(termDocumentMatrix, wordcount, minfreq = 2)
}
\arguments{
\item{termDocumentMatrix}{tm TermDocumentMatrix object}

\item{wordcount}{How many words to graph}

\item{minfreq}{minimum number of times a word should occur to be reported}

\item{shortendoc}{Should document names be shortened?}
}
\value{
a d3heatmap object.
}
\description{
This function creates a d3 interactive heatmap
}
\examples{
word_heatmap(TermDocumentMatrix(corpus1),10,minfreq=5)
}
\seealso{
\code{\link{d3heatmap}}
}

