% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{wordcount_table}
\alias{wordcount_table}
\title{Creates a table of word counts within a set of documents.}
\usage{
wordcount_table(wordlist, termDocumentMatrix, doccorpus, trunc = FALSE,
  raw = FALSE, onlywordlist = F)
}
\arguments{
\item{wordlist}{list of words}

\item{termDocumentMatrix}{tm TermDocumentMatrix object}

\item{doccorpus}{original document corpus}

\item{trunc}{should filenames be truncated in output table}

\item{raw}{if true then outputs a data.table if false then produces an htmltable allowing csv/excel export. the resulting table can be saved using the saveWidgets command in htmlwidgets.}

\item{onlywordlist}{if true, bins words, otherwise returns all unique matches}

\item{weighting}{passed to tm::TermDocumentMatrix (by default is weightTf)}
}
\value{
a data.table or DT datatable depending on if raw is set to true or false.
}
\description{
This function is useful for looking at occurences of words within documents
}
\examples{
wordcount_table(c("gender","nutrition"),TermDocumentMatrix(clcorp),corpus1)
}

