% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordtable-data.R
\docType{data}
\name{wordtable}
\alias{wordtable}
\title{Example EPAR data}
\format{An object of class \code{"tibble"}; see \code{\link[qtl]{tible:::tible}}.}
\source{
\href{sqlunet}{sqlunet}
}
\usage{
data(wordtable)
}
\description{
Data downloaded from the epar website.
}
\examples{
data(wordtable)
}
\references{
sqlunet
(\href{sqlunet.sourceforge.com}{Sourceforge})
}
\keyword{datasets}
