% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicFunctions.R
\name{writeFormulaforSTM}
\alias{writeFormulaforSTM}
\title{Write the formula for the call to the topic model.}
\usage{
writeFormulaforSTM(BASE_INPUT, workingfolder)
}
\arguments{
\item{BASE_INPUT}{the baseinput object returned by PreTopicFrame}

\item{workingfolder}{the workingfolder where you are saving objects.}
}
\value{
writes the formula to a text file in the working formula named formula1.
}
\description{
Write the formula for the prevelence covariates for the topic model.
}
\examples{
writeFormulaforSTM(BASE_INPUT,"../Research.Grants")
}
\seealso{
\code{\link{stm}}
}

