% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicFunctions.R
\name{writerun_alch}
\alias{writerun_alch}
\title{Function to write code to run alchemy}
\usage{
writerun_alch(alchemykey, workingfolder)
}
\arguments{
\item{alchemykey}{key to alchemy api as character}

\item{workingfolder}{workingfolder for project}
}
\value{
a .R file saved in the working folder
}
\description{
This function creates a .R file to call to alchemy api.
}

