% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunctions.R
\name{word_heatmap}
\alias{word_heatmap}
\title{TermDocumentMatrix to interactive heatmap.}
\usage{
word_heatmap(termDocumentMatrix, wordcount, minfreq = 2,
  col_labels = paste(1:ncol(termDocumentMatrix),
  stringr::str_sub(colnames(termDocumentMatrix), -10, -1)), dendrows = TRUE,
  dendcolumns = TRUE, pickwords = c())
}
\arguments{
\item{termDocumentMatrix}{tm TermDocumentMatrix object}

\item{wordcount}{How many words to graph}

\item{minfreq}{minimum number of times a word should occur to be reported}

\item{col_labels}{document labels in order of tdm columns}

\item{dendrows}{draw a dendogram for rows}

\item{dendcolumns}{draw a dendogram fo columns}

\item{pickwords}{list of words to be included in heatmap. Default is all words sorted by frequency.}
}
\value{
a d3heatmap object.
}
\description{
This function creates a d3 interactive heatmap
}
\examples{
word_heatmap(TermDocumentMatrix(corpus1),10)
}
\seealso{
\code{\link{d3heatmap}}
}

